uptime: move some codes to uucore
