+++ src/uu/uptime/Cargo.toml
@@ -19,17 +19,17 @@ path = "src/uptime.rs"
 [dependencies]
 chrono = { workspace = true }
 clap = { workspace = true }
-uucore = { workspace = true, features = ["libc", "utmpx"] }
 thiserror = { workspace = true }
-
-[target.'cfg(target_os="linux")'.dependencies]
-libsystemd-sys = { workspace = true }
+uucore = { workspace = true, features = ["libc", "utmpx", "uptime"] }
 
 [target.'cfg(target_os = "openbsd")'.dependencies]
 utmp-classic = { workspace = true }
 
 [target.'cfg(target_os="windows")'.dependencies]
-windows-sys = {workspace = true, features = ["Win32_System_RemoteDesktop", "Wdk_System_SystemInformation"]}
+windows-sys = { workspace = true, features = [
+  "Win32_System_RemoteDesktop",
+  "Wdk_System_SystemInformation",
+] }
 
 [[bin]]
 name = "uptime"
